requirements.txt
streamlit==1.35.0
fpdf==1.7.2
google-generativeai==0.4.1
pydantic==2.7.1
google-adk==0.0.5